kernel: wifi: mediatek: apply patches from wive-ng

+++ trunk/proprietary/rt_wifi/rtpci/3.0.X.X/mt76x2/include/smart_mesh/smart_mesh_def.h
@@ -274,7 +274,7 @@ typedef struct GNU_PACKED _SMART_RADIO_INFO{
 	CHAR MCS;
 	CHAR BW;
 	CHAR ShortGI;
-	UINT32 RATE;
+	ULONG RATE;
 } SMART_RADIO_INFO, *PSMART_RADIO_INFO;
 
 typedef struct GNU_PACKED _SMART_MESH_RAW {
@@ -366,10 +366,10 @@ typedef struct _SMART_MESH_CLI_INFO {
 	
 	/* statistics */
 	INT64	rx_total_packets;
-	ULONG	rx_total_bytes;
+	INT64	rx_total_bytes;
 	ULONG	rx_fail_cnt;
 	INT64	tx_total_packets;
-	ULONG	tx_total_bytes;
+	INT64	tx_total_bytes;
 	ULONG	tx_fail_cnt;
 } SMART_MESH_CLI_INFO, *PSMART_MESH_CLI_INFO;
